@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 0.1.3
 
 environment:
-  sdk: '>=3.4.0 <4.0.0'
+  sdk: '>=3.5.2 <4.0.0'
 
 dependencies:
   cupertino_icons: ^1.0.2
@@ -22,9 +22,9 @@ dependencies:
   json_annotation: ^4.8.1
   # Networking
   dio: ^5.3.3
-  retrofit: ^4.0.3
-  retrofit_generator: ^8.0.4
-  pretty_dio_logger: ^1.3.1
+  retrofit: ^4.3.0
+  retrofit_generator: ^9.0.0
+  pretty_dio_logger: ^1.4.0
   # localization
   intl: ^0.19.0
   easy_localization: ^3.0.7
@@ -45,13 +45,12 @@ dependencies:
   flutter:
     sdk: flutter
   bloc: ^8.1.4
-
+    
 
   injectable: ^2.4.2
   equatable: ^2.0.5
   animate_do: ^3.3.4
   vector_graphics: ^1.1.11+1
-  vector_graphics_compiler: ^1.1.11+1
   synchronized: ^3.1.0+1
   lottie: ^3.1.2
   loading_animation_widget: ^1.2.1
@@ -69,11 +68,14 @@ dependencies:
   path: ^1.9.0
   flutter_switch: ^0.3.2
   shimmer: ^3.0.0
-
+  showcaseview: ^3.0.0
+  flutter_slidable: ^3.1.1
+  flutter_contacts: ^1.1.9+2
 dev_dependencies:
+
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.0
+  flutter_lints: ^4.0.0
 
 flutter:
   fonts:
@@ -92,8 +94,6 @@ flutter:
     - assets/jsons/
     - assets/translation/en-US.json
     - assets/translation/ar-SA.json
-        # Apply transformer to a folder of SVGs
-    # - path: assets/images/asset 0.svg
     - path: assets/icons/
       transformers:
         - package: vector_graphics_compiler
