@@ -1,3 +1,4 @@
+import 'package:bulk_app/core/helpers/contacts_service.dart';
 import 'package:bulk_app/core/helpers/media.dart';
 import 'package:bulk_app/core/networking/api_service.dart';
 import 'package:bulk_app/core/networking/dio_factory.dart';
@@ -9,6 +10,7 @@ import 'package:bulk_app/features/manage_audiances/logic/manage_contact_cubit/co
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 import 'package:image_picker/image_picker.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../features/auth/data/repos/auth_repo.dart';
 import '../../features/auth/logic/login_cubit/login_cubit.dart';
@@ -21,9 +23,12 @@ final getIt = GetIt.instance;
 Future<void> setupGetIt() async {
   // Dio & ApiService
   Dio dio = DioFactory.getDio();
+  final sharedPrefs = await SharedPreferences.getInstance();
+  getIt.registerLazySingleton<SharedPreferences>(() => sharedPrefs);
+  // getIt.registerLazySingleton<FlutterContactPickerPlus>(
+  //     () => FlutterContactPickerPlus());
+  getIt.registerLazySingleton<ContactsService>(() => ContactsServiceImpl());
   getIt.registerLazySingleton<ApiService>(() => ApiService(dio));
-
-  // // login
 
   // Repos
   getIt.registerLazySingleton<AuthRepo>(() => AuthRepo(getIt()));
@@ -39,15 +44,10 @@ Future<void> setupGetIt() async {
   getIt.registerFactory<AddTemplateCubit>(() => AddTemplateCubit(getIt()));
   getIt.registerFactory<ManageAudiancesCubit>(
       () => ManageAudiancesCubit(getIt()));
-  getIt.registerFactory<ContactScreenCubit>(() => ContactScreenCubit(getIt()));
+  getIt.registerFactory<ContactScreenCubit>(
+      () => ContactScreenCubit(getIt(), getIt()));
+
   //App Media
   getIt.registerLazySingleton<AppMedia>(
       () => AppMedia(imagePicker: ImagePicker()));
-  // // signup
-  // getIt.registerLazySingleton<SignupRepo>(() => SignupRepo(getIt()));
-  // getIt.registerFactory<SignupCubit>(() => SignupCubit(getIt()));
-
-  // home
-  // getIt.registerLazySingleton<HomeApiService>(() => HomeApiService(dio));
-  // getIt.registerLazySingleton<HomeRepo>(() => HomeRepo(getIt()));
 }
