@@ -27,7 +27,7 @@ extension StringExtension on String? {
 
 extension ListExtension<T> on List<T>? {
   bool isNullOrEmpty() => this == null || this!.isEmpty;
-  bool isNotNullAndNotEmpty() => this != null || this!.isNotEmpty;
+  bool isNotNullAndNotEmpty() => this != null && this!.isNotEmpty;
 }
 
 extension SizedContext on BuildContext {
