@@ -1,4 +1,5 @@
 import 'package:bloc/bloc.dart';
+import 'package:bulk_app/core/helpers/contacts_service.dart';
 import 'package:bulk_app/core/networking/api_error_model.dart';
 import 'package:bulk_app/features/manage_audiances/data/models/audiance_response_data.dart';
 import 'package:bulk_app/features/manage_audiances/data/models/audience.dart';
@@ -16,24 +17,26 @@ part 'contact_screen_state.dart';
 class ContactScreenCubit extends Cubit<ContactScreenState> {
   late final AudienceRepository _audienceRepository;
 
-  ContactScreenCubit(this._audienceRepository)
+  ContactScreenCubit(this._audienceRepository, this._contactsService)
       : super(const ContactScreenState.initial());
 
   List<Contact>? currentAudienceContacts;
   Audience? currentAudience;
-  final audienceNameController = TextEditingController();
+  final audienceNameController = RestorableTextEditingController();
   final contactNameController = TextEditingController();
   final contactNumberController = TextEditingController();
-  // void init(Audiences audience) {
-  //   emit(const ContactScreenState.loading());
-  //   currentAudience = audience;
-  //   currentAudienceContacts = audience.contacts;
-  //   emit(ContactScreenState.loaded(currentAudienceContacts ?? []));
-  // }
-
+  bool isValid = false;
+  changeValidButton() {
+    if (audienceNameController.value.text.isNotEmpty) {
+      isValid = true;
+    } else {
+      isValid = false;
+    }
+    emit(ContactScreenState.isValidButtonState(isValid));
+  }
 // get audience by id
 
-  void getAudienceById(String id) async {
+  void emitGetAudienceByIdStates(String id) async {
     emit(const ContactScreenState.getContactsFromServerLoadingState());
 
     final response = await _audienceRepository.getAudienceById(id);
@@ -70,10 +73,28 @@ class ContactScreenCubit extends Cubit<ContactScreenState> {
     emit(ContactScreenState.contactAdded(currentAudienceContacts ?? []));
   }
 
+  void addContactsList(List<Contact> contacts) {
+    // Contact newContact = Contact(name: contact.name, phone: contact.phone);
+    // List<String> currentAudienceContactsNames =
+    //     currentAudienceContacts!.map((e) => e.name!).toList();
+    currentAudienceContacts = [
+      ...currentAudienceContacts ?? [],
+      ...contacts.where((item) => !currentAudienceContacts!
+          .map((e) => e.name)
+          .toList()
+          .contains(item.name))
+    ];
+
+    currentAudience?.contacts = currentAudienceContacts;
+
+    emit(ContactScreenState.contactAdded(currentAudienceContacts ?? []));
+  }
+
   void emitAddContactsToServerStates() async {
     emit(const ContactScreenState.addContactsToServerLoadingState());
     final response = await _audienceRepository.addNewAudience(Audience(
-        contacts: currentAudienceContacts, name: audienceNameController.text));
+        contacts: currentAudienceContacts,
+        name: audienceNameController.value.text));
     if (kDebugMode) {
       print(response);
     }
@@ -82,5 +103,54 @@ class ContactScreenCubit extends Cubit<ContactScreenState> {
     }, failure: (error) {
       emit(ContactScreenState.addContactsToServerErrorsState(error));
     });
+  }
+
+  void emitUpdateContactsToServerStates() async {
+    emit(const ContactScreenState.updateContactsInServerLoadingState());
+    final response = await _audienceRepository.updateAudience(Audience(
+      id: currentAudience?.id,
+      name: audienceNameController.value.text,
+      contacts: currentAudienceContacts,
+    ));
+    if (kDebugMode) {
+      print(response);
+    }
+    response.when(success: (response) async {
+      emit(const ContactScreenState.updateContactsInServerSuccessState());
+    }, failure: (error) {
+      emit(ContactScreenState.updateContactsInServerErrorState(error));
+    });
+  }
+
+  final ContactsService _contactsService;
+
+  Future<void> syncContacts() async {
+    // emit(
+    //   ContactsLoading(
+    //     contacts: state.contacts,
+    //   ),
+    // );
+    final contacts = await _contactsService.getContacts();
+    print('contactscontacts${contacts?.length}');
+    if (contacts != null) {
+      if (contacts.isNotEmpty) {
+        print(contacts);
+        // emit(ContactsSuccess(contacts: contacts));
+      } else {
+        // emit(
+        // ContactsFailure(
+        //   contacts: state.contacts,
+        //   errorMessage: 'No contacts were found on the device',
+        // ),
+        // );
+      }
+    } else {
+      // emit(
+      //   ContactsFailure(
+      //     contacts: state.contacts,
+      //     errorMessage: 'Unable to fetch contacts',
+      //   ),
+      // );
+    }
   }
 }
