@@ -3,6 +3,7 @@ import 'package:bulk_app/features/manage_audiances/ui/widgets/manage_audiances_w
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:showcaseview/showcaseview.dart';
 
 import '../../../../../core/resources/app_strings.dart';
 import '../../../../../core/theming/colors.dart';
@@ -13,22 +14,26 @@ class ManageAudiancesBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.symmetric(horizontal: 20.w, vertical: 20.h),
-      child: Column(
-        children: [
-          AppTextField(
-            hintText: AppStrings.searchAudiences.tr(),
-            prefixIcon: Icon(
-              Icons.search,
-              size: 25.sp,
-              color: ColorsManager.saerchTextFieldHintColor,
-            ),
+    return ShowCaseWidget(
+      builder: (context) {
+        return Padding(
+          padding: EdgeInsets.symmetric(horizontal: 20.w, vertical: 20.h),
+          child: Column(
+            children: [
+              AppTextField(
+                hintText: AppStrings.searchAudiences.tr(),
+                prefixIcon: Icon(
+                  Icons.search,
+                  size: 25.sp,
+                  color: ColorsManager.saerchTextFieldHintColor,
+                ),
+              ),
+              20.sizedBoxHeight,
+              const Expanded(child: AudiancesListView()),
+            ],
           ),
-          20.sizedBoxHeight,
-          const Expanded(child: AudiancesListView()),
-        ],
-      ),
+        );
+      }
     );
   }
 }
